chore(wren-ui): add common, repositories & adaptors to keep consistent format
