+++ wren-ui/src/apollo/server/repositories/index.ts
@@ -1,3 +1,4 @@
+export * from './baseRepository';
 export * from './learningRepository';
 export * from './modelRepository';
 export * from './projectRepository';
@@ -6,5 +7,8 @@ export * from './modelNestedColumnRepository';
 export * from './relationshipRepository';
 export * from './metricsRepository';
 export * from './metricsMeasureRepository';
+export * from './deployLogRepository';
 export * from './viewRepository';
-export * from './baseRepository';
+export * from './threadRepository';
+export * from './threadResponseRepository';
+export * from './schemaChangeRepository';
