+++ wren-ui/src/common.ts
@@ -1,0 +1,127 @@
+import { getConfig } from '@server/config';
+import { bootstrapKnex } from './apollo/server/utils/knex';
+import {
+  ProjectRepository,
+  ViewRepository,
+  DeployLogRepository,
+  ThreadRepository,
+  ThreadResponseRepository,
+  ModelRepository,
+  ModelColumnRepository,
+  RelationRepository,
+  SchemaChangeRepository,
+  ModelNestedColumnRepository,
+  LearningRepository,
+} from '@server/repositories';
+import {
+  WrenEngineAdaptor,
+  WrenAIAdaptor,
+  IbisAdaptor,
+} from '@server/adaptors';
+import {
+  DataSourceMetadataService,
+  QueryService,
+  ProjectService,
+  DeployService,
+  AskingService,
+} from '@server/services';
+import { PostHogTelemetry } from './apollo/server/telemetry/telemetry';
+
+export const serverConfig = getConfig();
+
+export const initComponents = () => {
+  const telemetry = new PostHogTelemetry();
+  const knex = bootstrapKnex({
+    dbType: serverConfig.dbType,
+    pgUrl: serverConfig.pgUrl,
+    debug: serverConfig.debug,
+    sqliteFile: serverConfig.sqliteFile,
+  });
+
+  // repositories
+  const projectRepository = new ProjectRepository(knex);
+  const deployLogRepository = new DeployLogRepository(knex);
+  const threadRepository = new ThreadRepository(knex);
+  const threadResponseRepository = new ThreadResponseRepository(knex);
+  const viewRepository = new ViewRepository(knex);
+  const modelRepository = new ModelRepository(knex);
+  const modelColumnRepository = new ModelColumnRepository(knex);
+  const modelNestedColumnRepository = new ModelNestedColumnRepository(knex);
+  const relationRepository = new RelationRepository(knex);
+  const schemaChangeRepository = new SchemaChangeRepository(knex);
+  const learningRepository = new LearningRepository(knex);
+
+  // adaptors
+  const wrenEngineAdaptor = new WrenEngineAdaptor({
+    wrenEngineEndpoint: serverConfig.wrenEngineEndpoint,
+  });
+  const wrenAIAdaptor = new WrenAIAdaptor({
+    wrenAIBaseEndpoint: serverConfig.wrenAIEndpoint,
+  });
+  const ibisAdaptor = new IbisAdaptor({
+    ibisServerEndpoint: serverConfig.ibisServerEndpoint,
+  });
+
+  // services
+  const metadataService = new DataSourceMetadataService({
+    ibisAdaptor,
+    wrenEngineAdaptor,
+  });
+  const projectService = new ProjectService({
+    projectRepository,
+    metadataService,
+  });
+  const queryService = new QueryService({
+    ibisAdaptor,
+    wrenEngineAdaptor,
+    telemetry,
+  });
+  const deployService = new DeployService({
+    wrenAIAdaptor,
+    deployLogRepository,
+    telemetry,
+  });
+  const askingService = new AskingService({
+    telemetry,
+    wrenAIAdaptor,
+    deployService,
+    projectService,
+    viewRepository,
+    threadRepository,
+    threadResponseRepository,
+    queryService,
+  });
+
+  return {
+    knex,
+    telemetry,
+
+    // repositories
+    projectRepository,
+    deployLogRepository,
+    threadRepository,
+    threadResponseRepository,
+    viewRepository,
+    modelRepository,
+    modelColumnRepository,
+    relationRepository,
+    schemaChangeRepository,
+    learningRepository,
+    modelNestedColumnRepository,
+
+    // adaptors
+    wrenEngineAdaptor,
+    wrenAIAdaptor,
+    ibisAdaptor,
+
+    // services
+    metadataService,
+    projectService,
+    queryService,
+    deployService,
+    askingService,
+  };
+};
+
+// singleton components
+export const components = initComponents();
